requirements.txt
pymongo==4.8.0
pandas==2.2.3
python-dotenv==1.0.1
tabulate==0.9.0
openpyxl